requirements-cpu.txt
tensorflow-intel
numpy==1.26.4
scikit-learn>=1.3.0
pillow>=10.0.0
opencv-python<4.10
matplotlib>=3.8.0
flask>=3.0.0
gunicorn>=21.2.0
